Adding commit id detection on interactive mode
If this extractor is used in interactive mode receiving the commit as a
single string, it expects the very first line in that string to contain
the revision number "r<NUMBER>" and uses it as the commit id. If this
line is not present, it falls back to "<no_id>" as commit id.